Upgrade NLog to 4.7.15 to support structured logging and start deployment of structured logging in Replay consumer. Add warning if the delay is skipped because processing is behind. Fix FileWorkloadListener count being wrong when applying filters. Fix FileWorkloadListener returning events with no event sequence first, which causes synchronized mode to not work because the start time could be any time within the replay. Fix FileWorkloadListener causing the Sqlite result set to slowly copy to a temp file due to no index in the Events table (index added in WorkloadFileWriterConsumer). Fix ReplayWorker threads may sleep for too long due to hardware thread-exhaustion on busy workloads. By creating the tasks as LongRunning thread over-subscription is allowed which works well when there are lots of threads that may be sleeping. Fix ReplayWorker wasting threads by ending the Task when the command queue is empty. ReplayConsumer will restart it when the next command is dequeued or it will be disposed when InactiveWorkerTerminationTimeoutSeconds is reached. Fix ReplayOffset is inconsistently a long or double. Enhance ReplayWorker to make publishing errors to sp_trace_generateevent optional to improve performance when it isn't required (RaiseErrorsToSqlEventTracing parameter). Enhance ReplayWorker to allow errors to output to NLog at a customisable level (CommandErrorLogLevel parameter). Enhance ReplayWorker to include the name of the session in every log message.

+++ WorkloadTools/ExecutionWorkloadEvent.cs
@@ -21,6 +21,6 @@ namespace WorkloadTools
         public long? EventSequence { get; set; }
         // This is the requested offset in milliseconds
         // from the the beginning of the workload
-        public long ReplayOffset { get; set; } = 0; // MILLISECONDS 
+        public double ReplayOffset { get; set; } = 0; // MILLISECONDS 
     }
 }
